requirements.txt
streamlit==1.38.0
yfinance==0.2.54
pandas==2.2.2
plotly==5.23.0
google-generativeai==0.8.4